requirements.txt
sentence-transformers>=3.4.1
faiss-cpu>=1.13.0
numpy>=2.2.0
matplotlib>=3.10.0
datasets>=3.5.0
torch>=2.6.0
tqdm>=4.67.0
tabulate>=0.9.0
